Encrypt a message using EVP functions.
Includes intilization for all the EVP contexts. We also generate the private key pair and the public session key in in the init function.

EVP is the AES prescribed crypto functions which include a combination of symmetric and symmetric ciphers.

This is a work in progress commit.

I will also implement the decrypt part and add UTs in later commits.